Move timeout up to the whole-host level
The previous timeout stuff didn't cover enough cases. This should
instead apply the timeout to handling the whole host which should
include connection/commands/disconnection.

This requires twisted 18.7 to ensure that when the per-host deferred is
cancelled we propagate the cancellation down through @inlineCallbacks to
the actual connections on the hosts.